add second plot
